+++ frontend/src/routes/login/+page.svelte
@@ -26,7 +26,6 @@
 
 		if (emailInput.value && passwordInput.value) {
 			const res = await authService.login(emailInput.value, passwordInput.value);
-			console.log(res);
 			alert(`LOGGED IN AS: ${res.username}`);
 		}
 	}
